refactor: split large modules into directory modules, dissolve common.rs
Reorganize the flat src/ layout (seven files over 500 lines) into
directory modules split by concern; every file is now under ~570 lines.
All moves done with git mv, no behavior changes.

- dissolve common.rs: Timing -> processor, retry_with_clients -> client,
  throttler -> stats, random primitives + payload naming -> generators::random
- generators/: PointGenerator trait, legacy (flag-driven) and config
  (YAML-driven) point generators, search query generator, random primitives
- config/: UploadConfig + load/validate in mod.rs; types split into
  collection/vector/payload; absorb search_config.rs and schema.rs
- collection/, search/: split along the existing from-args vs from-config seam,
  shared helpers (wait_index, recall scoring) in mod.rs
- args/: extract WriteOrdering/ReadConsistency value types into consistency.rs
- gitignore local benchmark data (tmp-data/, logs/, *.fbin, root *.jsonl)
- document the project layout in DEVELOPMENT.md
