+++ .gitignore
@@ -1,3 +1,9 @@
 /target
 .idea/
 *.iml
+
+# Local benchmark data & outputs
+/tmp-data/
+/logs/
+*.fbin
+/*.jsonl
