+++ src/config/mod.rs
@@ -6,6 +6,19 @@
 //!
 //! See `BENCHMARK_ROADMAP.md` §2 for the schema rationale and `examples/`.
 
+pub mod collection;
+pub mod payload;
+pub mod schema;
+pub mod search;
+pub mod vector;
+
+pub use collection::{CollectionConfig, IdType, QuantKind};
+pub use payload::{PayloadSource, PayloadSourceKind, PayloadType, TokenizerKind};
+pub use vector::{
+    ComparatorKind, DatatypeKind, DistanceKind, DistributionKind, FileStrategy, SparseKind,
+    SparseSource, VectorConfig, VectorSource,
+};
+
 use std::fmt;
 use std::marker::PhantomData;
 use std::path::Path;
@@ -15,405 +28,11 @@ use anyhow::{Context, Result, bail};
 use serde::de::{self, MapAccess, Visitor};
 use serde::{Deserialize, Deserializer, Serialize};
 
-use crate::dataset::DatasetConfig;
-
 /// Top-level document: `{ collection: { ... } }`.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
 pub struct UploadConfig {
     pub collection: CollectionConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct CollectionConfig {
-    #[serde(default = "default_collection_name")]
-    pub name: String,
-    #[serde(default)]
-    pub id: IdType,
-    #[serde(default = "default_true")]
-    pub on_disk_payload: bool,
-    pub shard_number: Option<u32>,
-    #[serde(default = "default_one")]
-    pub replication_factor: u32,
-    #[serde(default = "default_one")]
-    pub write_consistency_factor: u32,
-    pub sharding: Option<ShardingConfig>,
-    pub hnsw: Option<HnswConfig>,
-    pub optimizers: Option<OptimizersConfig>,
-    pub quantization: Option<QuantizationConfig>,
-    #[serde(default)]
-    pub vectors: Vec<VectorConfig>,
-    #[serde(default)]
-    pub sparse_vectors: Vec<SparseVectorConfig>,
-    /// Payload-wide settings, notably the whole-payload `source`.
-    #[serde(default)]
-    pub payload: PayloadSection,
-    /// Payload field declarations: value generation and/or which fields to index.
-    #[serde(default)]
-    pub fields: Vec<PayloadConfig>,
-}
-
-/// Payload-wide settings (`collection.payload`).
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct PayloadSection {
-    /// Whole-payload source: when set (to `type: dataset`), every point's entire
-    /// payload object is loaded from the dataset's `payloads.jsonl`. `fields`
-    /// entries then only need to declare which fields to index (they may omit
-    /// their own `source`); fields not listed are uploaded but left unindexed.
-    #[serde(default)]
-    pub source: Option<PayloadSource>,
-}
-
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "snake_case")]
-pub enum IdType {
-    #[default]
-    Integer,
-    Uuid,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct ShardingConfig {
-    #[serde(default = "default_custom")]
-    pub method: String,
-    pub key: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct HnswConfig {
-    pub m: Option<u64>,
-    pub payload_m: Option<u64>,
-    pub ef_construct: Option<u64>,
-    pub full_scan_threshold: Option<u64>,
-    #[serde(default)]
-    pub on_disk: bool,
-    #[serde(default)]
-    pub inline_storage: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct OptimizersConfig {
-    pub default_segment_number: Option<u64>,
-    pub indexing_threshold: Option<u64>,
-    pub memmap_threshold: Option<u64>,
-    pub max_segment_size: Option<u64>,
-    #[serde(default)]
-    pub prevent_unoptimized: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct QuantizationConfig {
-    #[serde(rename = "type")]
-    pub kind: QuantKind,
-    #[serde(default)]
-    pub always_ram: bool,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "kebab-case")]
-pub enum QuantKind {
-    None,
-    Scalar,
-    Binary,
-    #[serde(rename = "binary-2bit")]
-    Binary2bit,
-    #[serde(rename = "binary-1.5bit")]
-    Binary15bit,
-    #[serde(rename = "turbo-1bit")]
-    Turbo1bit,
-    #[serde(rename = "turbo-1.5bit")]
-    Turbo15bit,
-    #[serde(rename = "turbo-2bit")]
-    Turbo2bit,
-    #[serde(rename = "turbo-4bit")]
-    Turbo4bit,
-    #[serde(rename = "product-x4")]
-    ProductX4,
-    #[serde(rename = "product-x8")]
-    ProductX8,
-    #[serde(rename = "product-x16")]
-    ProductX16,
-    #[serde(rename = "product-x32")]
-    ProductX32,
-    #[serde(rename = "product-x64")]
-    ProductX64,
-}
-
-// ----------------------------- Dense vectors -----------------------------
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct VectorConfig {
-    /// Vector name. Omit for the unnamed default vector.
-    #[serde(default)]
-    pub name: Option<String>,
-    pub size: u64,
-    #[serde(default)]
-    pub distance: DistanceKind,
-    #[serde(default)]
-    pub datatype: DatatypeKind,
-    pub on_disk: Option<bool>,
-    pub multivector: Option<MultivectorConfig>,
-    pub quantization: Option<QuantizationConfig>,
-    #[serde(default, deserialize_with = "string_or_struct")]
-    pub source: VectorSource,
-}
-
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum DistanceKind {
-    #[default]
-    Cosine,
-    Dot,
-    Euclid,
-    Manhattan,
-}
-
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "snake_case")]
-pub enum DatatypeKind {
-    #[default]
-    Float32,
-    Float16,
-    Uint8,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct MultivectorConfig {
-    #[serde(default)]
-    pub comparator: ComparatorKind,
-    /// Number of sub-vectors to generate per point.
-    pub count: usize,
-}
-
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ComparatorKind {
-    #[default]
-    MaxSim,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "kebab-case", deny_unknown_fields)]
-pub enum VectorSource {
-    #[default]
-    Random,
-    File {
-        path: String,
-        #[serde(default)]
-        strategy: FileStrategy,
-    },
-    /// vector-db-benchmark dataset (specified inline in the source definition).
-    Dataset {
-        #[serde(flatten)]
-        dataset: DatasetConfig,
-    },
-}
-
-impl FromStr for VectorSource {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, String> {
-        match s {
-            "random" => Ok(VectorSource::Random),
-            other => Err(format!(
-                "unknown vector source {other:?}; expected \"random\" or a map with `type: file|dataset`"
-            )),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum FileStrategy {
-    #[default]
-    RandomSample,
-    FromStart,
-}
-
-// ----------------------------- Sparse vectors ----------------------------
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct SparseVectorConfig {
-    pub name: String,
-    #[serde(default)]
-    pub datatype: DatatypeKind,
-    #[serde(default)]
-    pub on_disk: bool,
-    #[serde(default, deserialize_with = "string_or_struct")]
-    pub source: SparseSource,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct SparseSource {
-    #[serde(default, rename = "type")]
-    pub kind: SparseKind,
-    #[serde(default = "default_vocab_size")]
-    pub vocab_size: usize,
-    #[serde(default = "default_sparse_length")]
-    pub length: usize,
-    #[serde(default)]
-    pub distribution: DistributionKind,
-    /// vector-db-benchmark dataset (specified inline under `dataset`).
-    #[serde(default)]
-    pub dataset: Option<DatasetConfig>,
-}
-
-impl Default for SparseSource {
-    fn default() -> Self {
-        SparseSource {
-            kind: SparseKind::Random,
-            vocab_size: default_vocab_size(),
-            length: default_sparse_length(),
-            distribution: DistributionKind::default(),
-            dataset: None,
-        }
-    }
-}
-
-impl FromStr for SparseSource {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, String> {
-        match s {
-            "random" => Ok(SparseSource::default()),
-            other => Err(format!(
-                "unknown sparse source {other:?}; expected \"random\""
-            )),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "kebab-case")]
-pub enum SparseKind {
-    #[default]
-    Random,
-    Dataset,
-}
-
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "kebab-case")]
-pub enum DistributionKind {
-    #[default]
-    Uniform,
-    Zipf,
-}
-
-// -------------------------------- Payloads -------------------------------
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct PayloadConfig {
-    pub name: String,
-    #[serde(rename = "type")]
-    pub kind: PayloadType,
-    /// Create a field index for this payload? `false` ⇒ unindexed filler.
-    #[serde(default = "default_true")]
-    pub index: bool,
-    #[serde(default)]
-    pub on_disk: bool,
-    #[serde(default)]
-    pub is_tenant: bool,
-    #[serde(default)]
-    pub is_principal: bool,
-    /// Integer payloads: also build a range index.
-    #[serde(default = "default_true")]
-    pub range_index: bool,
-    /// Text payloads: tokenizer.
-    pub tokenizer: Option<TokenizerKind>,
-    /// Per-field value source. May be omitted when `collection.payload.source`
-    /// provides the payload object — the entry is then index-only.
-    #[serde(default, deserialize_with = "option_string_or_struct")]
-    pub source: Option<PayloadSource>,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "snake_case")]
-pub enum PayloadType {
-    Keyword,
-    Integer,
-    Float,
-    Bool,
-    Uuid,
-    Geo,
-    Text,
-    Datetime,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum TokenizerKind {
-    Word,
-    Whitespace,
-    Prefix,
-    Multilingual,
-}
-
-/// All payload value-generation parameters. Which keys apply depends on the
-/// payload `type`; irrelevant keys are ignored.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct PayloadSource {
-    #[serde(default, rename = "type")]
-    pub kind: PayloadSourceKind,
-    /// vector-db-benchmark dataset for payload values (`type: dataset`).
-    #[serde(default)]
-    pub dataset: Option<DatasetConfig>,
-    /// Payload field name inside the dataset schema / `payloads.jsonl`.
-    #[serde(default)]
-    pub field: Option<String>,
-    // keyword
-    pub cardinality: Option<usize>,
-    pub length_multiplier: Option<usize>,
-    // keyword / integer: multivalue
-    pub values_per_point: Option<usize>,
-    // integer / float / datetime range
-    pub min: Option<f64>,
-    pub max: Option<f64>,
-    // bool
-    pub true_ratio: Option<f64>,
-    // geo
-    pub clusters: Option<usize>,
-    // text
-    pub vocab_size: Option<usize>,
-    pub min_length: Option<usize>,
-    pub max_length: Option<usize>,
-    #[serde(default)]
-    pub distribution: DistributionKind,
-}
-
-impl FromStr for PayloadSource {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, String> {
-        let kind = match s {
-            "random" => PayloadSourceKind::Random,
-            "random-clusters" => PayloadSourceKind::RandomClusters,
-            "now" => PayloadSourceKind::Now,
-            other => return Err(format!("unknown payload source {other:?}")),
-        };
-        Ok(PayloadSource {
-            kind,
-            ..Default::default()
-        })
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "kebab-case")]
-pub enum PayloadSourceKind {
-    #[default]
-    Random,
-    RandomClusters,
-    Now,
-    Dataset,
 }
 
 // ------------------------------ Loading / validation ---------------------
@@ -641,12 +260,6 @@ pub(crate) fn default_collection_name() -> String {
 }
 fn default_custom() -> String {
     "custom".to_string()
-}
-fn default_vocab_size() -> usize {
-    1000
-}
-fn default_sparse_length() -> usize {
-    100
 }
 
 #[cfg(test)]
